Create block_pyramids.c
Ever played Mario!! This program recreate those pyramids we seen in mario game using C, albeit in text, using hashes (#) for bricks. Each hash is a bit taller than it is wide, so the pyramids themselves are also be taller than they are wide.
here given the limit of height to 8, but you can change that.