Create readability.c
This is a cs50 question for computing the approximate grade level needed to comprehend some text. The readability test used is the Coleman-Liau index. The Coleman-Liau index of a text is designed to output what (U.S.) grade level is needed to understand the text. The formula is:
index = 0.0588 * L - 0.296 * S - 15.8